pj_project/requirements.txt
torch
transformers==4.37.2
bitsandbytes==0.41.3 
accelerate==0.25.0
Pillow==8.4.0
fastapi
uvicorn